requirements.txt
anyio==3.5.0
aplus==0.11.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astropy==5.0.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.1.0
blake3==0.3.1
bleach==4.1.0
bqplot==0.12.32
Brotli==1.0.9
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
dash==2.1.0
dash-bootstrap-components==1.0.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-pivottable==0.0.2
dash-table==5.0.0
dask==2022.1.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.2
fastapi==0.73.0
filelock==3.4.2
Flask==2.0.2
Flask-Assets==2.0
Flask-Caching==1.10.1
Flask-Compress==1.10.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
fonttools==4.29.1
frozendict==2.3.0
fsspec==2022.1.0
future==0.18.2
greenlet==1.1.2
h11==0.13.0
h5py==3.6.0
httptools==0.3.0
idna==3.3
importlib-resources==5.4.0
ipydatawidgets==4.2.0
ipykernel==6.9.0
ipyleaflet==0.15.0
ipympl==0.8.7
ipython==8.0.1
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipyvue==1.7.0
ipyvuetify==1.8.2
ipywebrtc==0.6.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
llvmlite==0.38.0
locket==0.2.1
MarkupSafe==2.0.1
matlab==0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.8
numba==0.55.1
numpy==1.21.5
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
plotly==5.6.0
progressbar2==4.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.27
pure-eval==0.2.2
pyarrow==7.0.0
pycparser==2.21
pydantic==1.9.0
pyerfa==2.0.0.1
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-utils==3.1.0
pythreejs==2.3.0
pytz==2021.3
pywin32==303
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
rich==11.2.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
stack-data==0.1.4
starlette==0.17.1
tabulate==0.8.9
tenacity==8.0.1
terminado==0.13.1
testpath==0.5.0
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
traittypes==0.2.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.4
vaex==4.8.0
vaex-astro==0.9.0
vaex-core==4.8.0
vaex-hdf5==0.12.0
vaex-jupyter==0.7.0
vaex-ml==0.17.0
vaex-server==0.8.1
vaex-viz==0.5.1
watchgod==0.7
wcwidth==0.2.5
webassets==2.0
webencodings==0.5.1
websockets==10.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wincertstore==0.2
wordcloud==1.8.1
xarray==0.21.1
xyzservices==2022.2.0
zipp==3.7.0
